+++ src/main/java/roomescape/reservation/service/ReservationQueryService.java
@@ -6,6 +6,7 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.global.exception.NotFoundException;
 import roomescape.reservation.controller.response.MyReservationResponse;
 import roomescape.reservation.controller.response.ReservationResponse;
 import roomescape.reservation.domain.Reservation;
@@ -18,8 +19,6 @@ import roomescape.waiting.service.WaitingQueryService;
 public class ReservationQueryService {
 
     private final ReservationRepository reservationRepository;
-    private final WaitingQueryService waitingQueryService;
-
     public List<ReservationResponse> getFilteredReservations(Long themeId, Long memberId, LocalDate from,
                                                              LocalDate to) {
         if (themeId == null && memberId == null && from == null && to == null) {
@@ -31,22 +30,18 @@ public class ReservationQueryService {
         return ReservationResponse.from(reservations);
     }
 
+    public Reservation getReservation(Long id) {
+        return reservationRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("예약을 찾을 수 없습니다."));
+    }
+
     private List<ReservationResponse> getAllReservations() {
         List<Reservation> reservations = reservationRepository.findAll();
 
         return ReservationResponse.from(reservations);
     }
 
-    public List<MyReservationResponse> getAllReservations(Long memberId) {
-        List<MyReservationResponse> responses = new ArrayList<>();
-
-        responses.addAll(getReservations(memberId));
-        responses.addAll(waitingQueryService.getWaitingReservations(memberId));
-
-        return responses;
-    }
-
-    private List<MyReservationResponse> getReservations(Long memberId) {
+    public List<MyReservationResponse> getReservations(Long memberId) {
         List<Reservation> reservations = reservationRepository.findByMemberId(memberId);
 
         return MyReservationResponse.from(reservations);
