+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -1,18 +1,21 @@
 package roomescape.reservation.service;
 
 import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.auth.web.exception.NotAuthorizationException;
 import roomescape.global.exception.InvalidArgumentException;
-import roomescape.global.exception.NotFoundException;
+import roomescape.reservation.controller.response.MyReservationResponse;
 import roomescape.reservation.controller.response.ReservationResponse;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservation.service.command.ReserveCommand;
 import roomescape.waiting.domain.Waiting;
 import roomescape.waiting.service.WaitingManager;
+import roomescape.waiting.service.WaitingQueryService;
 
 @RequiredArgsConstructor
 @Service
@@ -21,6 +24,8 @@ public class ReservationService {
     private final ReservationRepository reservationRepository;
     private final WaitingManager waitingManager;
     private final ReservationManager reservationManager;
+    private final ReservationQueryService reservationQueryService;
+    private final WaitingQueryService waitingQueryService;
 
     @Transactional
     public ReservationResponse reserve(ReserveCommand reserveCommand) {
@@ -43,11 +48,9 @@ public class ReservationService {
 
     @Transactional
     public void deleteByUser(Long id, Long userId) {
-        if (!reservationRepository.existsById(id)) {
-            throw new NotFoundException("예약을 찾을 수 없습니다.");
-        }
+        Reservation reservation = reservationQueryService.getReservation(id);
 
-        if (!reservationRepository.existsByIdAndMemberId(id, userId)) {
+        if (!reservation.isOwner(userId)) {
             throw new NotAuthorizationException("해당 예약자가 아닙니다.");
         }
 
@@ -56,7 +59,7 @@ public class ReservationService {
 
     @Transactional
     public void delete(Long id) {
-        Reservation reservation = getReservation(id);
+        Reservation reservation = reservationQueryService.getReservation(id);
         reservationRepository.delete(reservation);
         waitingToReservation(reservation);
     }
@@ -72,8 +75,14 @@ public class ReservationService {
         }
     }
 
-    private Reservation getReservation(Long id) {
-        return reservationRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("예약을 찾을 수 없습니다."));
+    @Transactional(readOnly = true)
+    public List<MyReservationResponse> getAllReservations(Long memberId) {
+        List<MyReservationResponse> responses = new ArrayList<>();
+
+        responses.addAll(reservationQueryService.getReservations(memberId));
+        responses.addAll(waitingQueryService.getWaitingReservations(memberId));
+
+        return responses;
     }
+
 }
